revert changes to generator and reuse logic in builder
