+++ bin/generate.dart
@@ -1,0 +1,5 @@
+import 'package:environment_config/generator.dart';
+
+void main(List<String> arguments) async {
+  await generateConfig(arguments);
+}
